requirements.txt
py-august
twilio
pyttsx3
SpeechRecognition
PyAudio